backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# API & Validation
pydantic>=2.7.4,<3.0.0
email-validator==2.3.0
pydantic-settings>=2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# AI/ML & Vector Search (minimal for Gemini + Pinecone only)
google-generativeai==0.8.0
pinecone-client==3.0.3
# sentence-transformers==2.2.2  # Optional: Only needed for Pinecone semantic search (900MB+ with PyTorch)

# PDF Processing
PyMuPDF==1.24.7

# Utilities
httpx==0.27.0
psutil==7.1.0
# redis==6.4.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3